Solve cosine-similarity

+++ cosine-similarity/cosine-similarity.py
@@ -1,8 +1,8 @@
 import numpy as np
 
 def cosine_similarity(a: list, b: list) -> float:
-    a = np.array(a, dtype=float)
-    b = np.array(b, dtype=float)
+    a = np.array(a)
+    b = np.array(b)
 
     norm_a = np.linalg.norm(a)
     norm_b = np.linalg.norm(b)
